+++ src/solidlsp/language_servers/clojure_lsp.py
@@ -6,7 +6,6 @@ import logging
 import os
 import pathlib
 import shutil
-import stat
 import subprocess
 import threading
 
@@ -118,7 +117,7 @@ class ClojureLSP(SolidLanguageServer):
             deps.install(logger, clojurelsp_ls_dir)
         if not os.path.exists(clojurelsp_executable_path):
             raise FileNotFoundError(f"Download failed? Could not find clojure-lsp executable at {clojurelsp_executable_path}")
-        os.chmod(clojurelsp_executable_path, stat.S_IEXEC)
+        os.chmod(clojurelsp_executable_path, 0o755)
         return clojurelsp_executable_path
 
     @staticmethod