+++ src/solidlsp/language_servers/omnisharp.py
@@ -6,7 +6,6 @@ import json
 import logging
 import os
 import pathlib
-import stat
 import threading
 from collections.abc import Iterable
 
@@ -194,7 +193,7 @@ class OmniSharp(SolidLanguageServer):
             FileUtils.download_and_extract_archive(logger, runtime_dependencies["OmniSharp"]["url"], omnisharp_ls_dir, "zip")
         omnisharp_executable_path = os.path.join(omnisharp_ls_dir, runtime_dependencies["OmniSharp"]["binaryName"])
         assert os.path.exists(omnisharp_executable_path)
-        os.chmod(omnisharp_executable_path, stat.S_IEXEC)
+        os.chmod(omnisharp_executable_path, 0o755)
 
         razor_omnisharp_ls_dir = os.path.join(cls.ls_resources_dir(solidlsp_settings), "RazorOmnisharp")
         if not os.path.exists(razor_omnisharp_ls_dir):