+++ src/solidlsp/ls_config.py
@@ -43,6 +43,7 @@ class Language(str, Enum):
     ELIXIR = "elixir"
     TERRAFORM = "terraform"
     SWIFT = "swift"
+    BASH = "bash"
     # Experimental or deprecated Language Servers
     TYPESCRIPT_VTS = "typescript_vts"
     """Use the typescript language server through the natively bundled vscode extension via https://github.com/yioneko/vtsls"""
@@ -106,6 +107,8 @@ class Language(str, Enum):
                 return FilenameMatcher("*.tf", "*.tfvars", "*.tfstate")
             case self.SWIFT:
                 return FilenameMatcher("*.swift")
+            case self.BASH:
+                return FilenameMatcher("*.sh", "*.bash")
             case _:
                 raise ValueError(f"Unhandled language: {self}")